Restructure and add component folders, create CSS files ffor each component, update import paths, fix capitalization on NewListButton component
