+++ src/components/Nav/Nav.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
+import './Nav.css';
 
 export default function Nav() {
   return (
